requirements.txt
soupsieve==2.3.2.post1
psycopg2-binary==2.9.5
aiobotocore==2.4.2
aiohttp==3.8.3
aioitertools==0.11.0
aiosignal==1.3.1
alembic==1.10.4
async-timeout==4.0.2
attrs==22.2.0
awscli==1.27.50
bcrypt==4.0.1
beautifulsoup4==4.11.1
blinker==1.6.2
boto==2.49.0
boto3==1.26.50
botocore==1.29.50
cachelib==0.9.0
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cloudpickle==2.2.1
colorama==0.4.4
contextlib2==21.6.0
cryptography==40.0.2
Cython==0.29.33
dash==2.9.3
dash-bootstrap-components==1.3.0
dash-core-components==2.0.0
dash-extensions==1.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dill==0.3.6
distro==1.8.0
dnspython==2.3.0
docker==6.0.1
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
docutils==0.16
EditorConfig==0.12.3
email-validator==2.0.0.post2
Flask==2.2.2
Flask-Caching==2.0.2
Flask-Dance==6.2.0
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.3
Flask-User==1.0.2.2
Flask-WTF==1.1.1
frozenlist==1.3.3
fsspec==2022.11.0
google-api-core==2.11.0
google-api-python-client==2.86.0
google-auth==2.17.3
google-auth-httplib2==0.1.0
google-pasta==0.2.0
googleapis-common-protos==1.59.0
httplib2==0.22.0
idna==3.4
importlib-metadata==4.13.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
jsbeautifier==1.14.7
jsonschema==3.2.0
Mako==1.2.4
MarkupSafe==2.1.1
more-itertools==9.1.0
multidict==6.0.4
multiprocess==0.70.14
numpy==1.24.1
oauthlib==3.2.2
packaging==23.0
pandas==1.5.2
paramiko==3.1.0
passlib==1.7.4
pathos==0.3.0
platformdirs==3.5.0
plotly==5.12.0
pox==0.3.2
ppft==1.7.6.6
protobuf==3.20.3
protobuf3-to-dict==0.1.5
pyasn1==0.4.8
pyasn1-modules==0.3.0
pycparser==2.21
PyNaCl==1.5.0
pyOpenSSL==23.1.1
pyparsing==3.0.9
pyrsistent==0.19.3
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==0.21.1
pytz==2022.7.1
PyYAML==5.4.1
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.7.2
s3fs==2022.11.0
s3transfer==0.6.0
sagemaker==2.151.0
schema==0.7.5
scikit-learn==1.2.0
scipy==1.10.0
six==1.16.0
smdebug-rulesconfig==1.0.1
SQLAlchemy==2.0.12
stripe==5.4.0
tblib==1.7.0
tenacity==8.1.0
texttable==1.6.7
threadpoolctl==3.1.0
typing_extensions==4.5.0
uritemplate==4.1.1
urllib3==1.26.8
URLObject==2.4.3
waitress==2.1.2
watchtower==3.0.1
websocket-client==0.59.0
Werkzeug==2.2.2
wrapt==1.14.1
WTForms==3.0.1
yarl==1.8.2
zipp==3.11.0
